+++ webmvc/src/main/java/net/brightroom/featureflag/configuration/FeatureFlagInterceptorRegistrationRule.java
